% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiva_serie.R
\name{descriptiva_serie}
\alias{descriptiva_serie}
\title{Analisis descriptivo a serie de tiempo}
\usage{
descriptiva_serie(
  data,
  nombre_serie = "Serie",
  titulo_grafico = nombre_serie,
  solo_grafico = FALSE
)
}
\arguments{
\item{data}{Serie de tiempo, debe ser ingresada como vector. Se recomienda usar as.vector() para ingresar la serie.}

\item{nombre_serie}{Nombre de la serie de tiempo. Si se deja vacio entonces se llamara "Serie". Se ingresa como cadena de caracteres.}

\item{titulo_grafico}{Titulo del grafico principal de la serie. Si se deja vacio entonces se toma el mismo nombre de la serie. Se ingresa como cadena de caracteres.}

\item{solo_grafico}{TRUE o FALSE, determina si quiere verse solo el grafico general o si tambien se imprime la informacion descriptiva. TRUE imprime solamente el grafico, FALSE imprime el grafico y la informacion descriptiva.}
}
\value{
Graficos y resultados
}
\description{
Esta funcion realiza el analisis descriptivo a una serie de tiempo, mostrando: Grafico de la serie, Tendencia, Media, Cuartiles, Prueba de normalidad, Prueba ruido blanco y Grafico distribucion de los datos.
}
\examples{
descriptiva_serie(data = as.vector(iris$Sepal.Length),
                  nombre_serie = 'Largo del Sepalo',
                  titulo_grafico = 'Longitud Sepalo')
}
